cleaning up the BE classes a bit
