+++ BGlib/be/dataset/__init__.py
@@ -1,6 +1,4 @@
-from . import be_raw_dataset
-from . import be_sho_dataset
-from . import ckpfm_dataset
+from . import be_sho_fit_datasets
 from . import be_raw_datasets
 
-__all__ = ['be_raw_dataset', 'be_sho_dataset', 'be_raw_datasets']
+__all__ = ['be_raw_datasets', 'be_sho_fit_datasets']
